# wikify-web/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Wikify Web API",
    "description": "AI-powered repository documentation and chat system",
    "contact": {
      "name": "Wikify Team",
      "email": "support@wikify.dev"
    },
    "license": {
      "name": "MIT OR Apache-2.0"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development server"
    },
    {
      "url": "https://api.wikify.dev",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/api/files/content": {
      "post": {
        "tags": [
          "Files"
        ],
        "summary": "Get file content",
        "operationId": "get_file_content",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetFileContentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "File content retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileContentResponse"
                }
              }
            }
          },
          "404": {
            "description": "File not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/files/readme": {
      "post": {
        "tags": [
          "Files"
        ],
        "summary": "Get README content for repository",
        "operationId": "get_readme",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetReadmeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "README content retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadmeResponse"
                }
              }
            }
          },
          "404": {
            "description": "Repository not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/files/tree": {
      "post": {
        "tags": [
          "Files"
        ],
        "summary": "Get file tree for repository",
        "operationId": "get_file_tree",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetFileTreeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "File tree retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileTreeResponse"
                }
              }
            }
          },
          "404": {
            "description": "Repository not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FileContentResponse": {
        "type": "object",
        "description": "Response containing file content",
        "required": [
          "repository_id",
          "file_path",
          "content",
          "size",
          "encoding"
        ],
        "properties": {
          "branch": {
            "type": [
              "string",
              "null"
            ]
          },
          "content": {
            "type": "string"
          },
          "encoding": {
            "type": "string"
          },
          "file_path": {
            "type": "string"
          },
          "repository_id": {
            "type": "string"
          },
          "size": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
      "FileTreeResponse": {
        "type": "object",
        "description": "Response containing file tree",
        "required": [
          "repository_id",
          "files",
          "total_files"
        ],
        "properties": {
          "branch": {
            "type": [
              "string",
              "null"
            ]
          },
          "files": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RepositoryFileInfo"
            }
          },
          "repository_id": {
            "type": "string"
          },
          "total_files": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
...
      "GetFileContentRequest": {
        "type": "object",
        "description": "Request to get file content",
        "required": [
          "repository_id",
          "file_path"
        ],
        "properties": {
          "branch": {
            "type": [
              "string",
              "null"
            ]
          },
          "file_path": {
            "type": "string"
          },
          "repository_id": {
            "type": "string"
          }
        }
      },
      "GetFileTreeRequest": {
        "type": "object",
        "description": "Request to get file tree",
        "required": [
          "repository_id"
        ],
        "properties": {
          "branch": {
            "type": [
              "string",
              "null"
            ]
          },
          "repository_id": {
            "type": "string"
          }
        }
      },
      "GetReadmeRequest": {
        "type": "object",
        "description": "Request to get README content",
        "required": [
          "repository_id"
        ],
        "properties": {
          "branch": {
            "type": [
              "string",
              "null"
            ]
          },
          "repository_id": {
            "type": "string"
          }
        }
      },
...
      "ReadmeResponse": {
        "type": "object",
        "description": "Response containing README content",
        "required": [
          "repository_id",
          "found"
        ],
        "properties": {
          "branch": {
            "type": [
              "string",
              "null"
            ]
          },
          "content": {
            "type": [
              "string",
              "null"
            ]
          },
          "found": {
            "type": "boolean"
          },
          "repository_id": {
            "type": "string"
          }
        }
      },
...
      "RepositoryFileInfo": {
        "type": "object",
        "description": "File information in the repository",
        "required": [
          "path",
          "file_type"
        ],
        "properties": {
          "file_type": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "sha": {
            "type": [
              "string",
              "null"
            ]
          },
          "size": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "minimum": 0
          }
        }
      },
...
    },
    "securitySchemes": {
      "api_key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  },
...
}